Moved osmosis test tube
