+++ contracts/counter/Cargo.toml
@@ -23,6 +23,7 @@ thiserror = { version = "1.0.21" }
 serde = { workspace = true }
 serde_json = "1.0.79"
 cw-orch = { path = "../../cw-orch", features = ["daemon"] }
+cw-orch-osmosis-test-tube = { version = "0.1.0", path = "../../packages/cw-orch-osmosis-test-tube" }
 
 [[example]]
 name = "deploy"
@@ -31,5 +32,5 @@ name = "deploy"
 # Deps for deployment
 dotenv = { version = "0.15.0" }
 pretty_env_logger = { version = "0.5.0" }
-cw-orch = { path = "../../cw-orch", features = ["daemon", "osmosis-test-tube"] }
+cw-orch = { path = "../../cw-orch", features = ["daemon"] }
 anyhow.workspace = true
