+++ contracts/counter/tests/osmosis-test-tube.rs
@@ -4,9 +4,9 @@ use counter_contract::{
     msg::{GetCountResponse, InstantiateMsg, QueryMsg},
     CounterContract,
 };
-use cw_orch::osmosis_test_tube::OsmosisTestTube;
 // Use prelude to get all the necessary imports
 use cw_orch::prelude::*;
+use cw_orch_osmosis_test_tube::OsmosisTestTube;
 
 // ANCHOR: integration_test
 // ANCHOR: setup
