+++ cw-orch/Cargo.toml
@@ -28,14 +28,6 @@ name = "testnet_daemon"
 required-features = ["daemon"]
 
 [[example]]
-name = "osmosis_test_tube"
-required-features = ["osmosis-test-tube"]
-
-[[example]]
-name = "complex_osmosis_test_tube"
-required-features = ["osmosis-test-tube"]
-
-[[example]]
 name = "injective"
 required-features = ["eth"]
 
@@ -60,12 +52,6 @@ daemon = [
   "dep:cw-orch-networks",
 ]
 eth = ["daemon", "cw-orch-core/eth", "cw-orch-daemon?/eth"]
-osmosis-test-tube = [
-  "dep:osmosis-test-tube",
-  "dep:prost",
-  "dep:prost-types",
-  "dep:cosmrs",
-]
 snapshot-testing = ["dep:insta", "dep:sanitize-filename"]
 
 [dependencies]
@@ -79,7 +65,7 @@ cw-orch-fns-derive = { workspace = true }
 # Daemon deps
 cw-orch-networks = { workspace = true, optional = true }
 tokio = { version = "1.4", features = ["full"], optional = true }
-cosmrs = { version = "0.15.0", features = [
+cosmrs = { workspace = true, features = [
   "dev",
   "cosmwasm",
   "grpc",
@@ -88,10 +74,6 @@ cosmrs = { version = "0.15.0", features = [
 # Live mock deps
 tonic = { workspace = true, optional = true, features = ["tls", "tls-roots"] }
 
-# Test Tube env deps
-osmosis-test-tube = { workspace = true, optional = true }
-prost-types = { workspace = true, optional = true }
-prost = { workspace = true, optional = true }
 
 # Snapshot deps
 insta = { version = "1.34.0", features = ["yaml"], optional = true }
@@ -123,3 +105,6 @@ mock-contract = { path = "../contracts/mock_contract" }
 mock-contract-u64 = { path = "../contracts/mock_contract_u64" }
 serde_json = { workspace = true }
 dotenv = "0.15.0"
+osmosis-std = "0.24.0"
+prost.workspace = true
+prost-types.workspace = true
