+++ cw-orch/src/prelude.rs
@@ -36,10 +36,6 @@ pub use crate::environment::{ChainInfo, ChainInfoOwned};
 // Mock for testing
 pub use crate::mock::{Mock, MockBech32};
 
-// OsmosisTestTube for testing
-#[cfg(feature = "osmosis-test-tube")]
-pub use crate::osmosis_test_tube::OsmosisTestTube;
-
 // error
 pub use crate::error::CwOrchError;
