+++ packages/cw-orch-osmosis-test-tube/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "cw-orch-osmosis-test-tube"
+version = "0.1.0"
+authors.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+
+# Test Tube env deps
+osmosis-test-tube = { workspace = true }
+prost-types = { workspace = true }
+prost = { workspace = true }
+cosmwasm-std.workspace = true
+serde.workspace = true
+cw-orch = { workspace = true }
+
+[dev-dependencies]
+counter-contract = { path = "../../contracts/counter" }
+env_logger = "0.11.3"
