+++ packages/cw-orch-osmosis-test-tube/src/lib.rs
@@ -1,6 +1,5 @@
 //! Integration testing execution environment backed by a [osmosis-test-tube](osmosis_test_tube) App.
 //! It has an associated state that stores deployment information for easy retrieval and contract interactions.
-
 mod core;
 
 mod queriers;
