+++ packages/cw-orch-osmosis-test-tube/src/queriers/wasm.rs
@@ -1,18 +1,18 @@
 use std::{cell::RefCell, marker::PhantomData, rc::Rc, str::FromStr};
 
-use cosmrs::AccountId;
 use cosmwasm_std::{
     from_json, instantiate2_address, to_json_vec, CanonicalAddr, CodeInfoResponse,
     ContractInfoResponse, HexBinary,
 };
-use cw_orch_core::{
+use cw_orch::{
     contract::interface_traits::Uploadable,
     environment::{Querier, QuerierGetter, StateInterface, WasmQuerier},
-    CwEnvError,
+    prelude::CwOrchError,
 };
+use osmosis_test_tube::cosmrs::AccountId;
 use osmosis_test_tube::{OsmosisTestApp, Runner};
 
-use crate::osmosis_test_tube::{map_err, OsmosisTestTube, MOCK_CHAIN_INFO};
+use crate::{map_err, OsmosisTestTube, MOCK_CHAIN_INFO};
 use osmosis_test_tube::osmosis_std::types::cosmwasm::wasm::v1::{
     QueryCodeRequest, QueryCodeResponse, QueryContractInfoRequest, QueryContractInfoResponse,
     QueryRawContractStateRequest, QueryRawContractStateResponse, QuerySmartContractStateRequest,
@@ -34,7 +34,7 @@ impl<S: StateInterface> OsmosisTestTubeWasmQuerier<S> {
 }
 
 impl<S> Querier for OsmosisTestTubeWasmQuerier<S> {
-    type Error = CwEnvError;
+    type Error = CwOrchError;
 }
 
 impl<S: StateInterface> QuerierGetter<OsmosisTestTubeWasmQuerier<S>> for OsmosisTestTube<S> {
@@ -57,7 +57,7 @@ impl<S: StateInterface> WasmQuerier for OsmosisTestTubeWasmQuerier<S> {
 
         Ok(code_info_result
             .code_info
-            .ok_or(CwEnvError::CodeIdNotInStore(code_id.to_string()))?
+            .ok_or(CwOrchError::CodeIdNotInStore(code_id.to_string()))?
             .data_hash
             .into())
     }
@@ -65,7 +65,7 @@ impl<S: StateInterface> WasmQuerier for OsmosisTestTubeWasmQuerier<S> {
     fn contract_info(
         &self,
         address: impl Into<String>,
-    ) -> Result<ContractInfoResponse, CwEnvError> {
+    ) -> Result<ContractInfoResponse, CwOrchError> {
         let address = address.into();
         let result = self
             .app
@@ -78,7 +78,7 @@ impl<S: StateInterface> WasmQuerier for OsmosisTestTubeWasmQuerier<S> {
             )
             .map_err(map_err)?
             .contract_info
-            .ok_or(CwEnvError::AddrNotInStore(address))?;
+            .ok_or(CwOrchError::AddrNotInStore(address))?;
 
         let mut contract_info = ContractInfoResponse::default();
         contract_info.code_id = result.code_id;
@@ -150,7 +150,7 @@ impl<S: StateInterface> WasmQuerier for OsmosisTestTubeWasmQuerier<S> {
 
         let code_info = response
             .code_info
-            .ok_or(CwEnvError::CodeIdNotInStore(code_id.to_string()))?;
+            .ok_or(CwOrchError::CodeIdNotInStore(code_id.to_string()))?;
 
         let mut c = CodeInfoResponse::default();
         c.code_id = code_id;
@@ -179,12 +179,12 @@ impl<S: StateInterface> WasmQuerier for OsmosisTestTubeWasmQuerier<S> {
     }
 
     fn local_hash<
-        T: cw_orch_core::contract::interface_traits::Uploadable
-            + cw_orch_core::contract::interface_traits::ContractInstance<Self::Chain>,
+        T: cw_orch::contract::interface_traits::Uploadable
+            + cw_orch::contract::interface_traits::ContractInstance<Self::Chain>,
     >(
         &self,
         _contract: &T,
-    ) -> Result<HexBinary, CwEnvError> {
+    ) -> Result<HexBinary, CwOrchError> {
         <T as Uploadable>::wasm(&MOCK_CHAIN_INFO.into()).checksum()
     }
 }
